move parser, create cli-layer for console execution
