+++ README.md
@@ -10,7 +10,9 @@ The application allows controlling robotic mowers on a rectangular plateau, proc
 ## Architecture
 - **model-layer**: Domain model (value objects, business logic)
 - **service-layer**: Application logic (MowerEngine, validation, handlers)
+- **adapter-layer**: Parsers/adapters for input/output (used by web and CLI)
 - **web-layer**: REST API (Spring Boot, controllers, OpenAPI/Swagger)
+- **cli-layer**: Command-line interface for interactive use
 - **docs/**: PlantUML diagrams for each layer
 
 > **Note:** Persistence is not supported, as it was not required in the problem statement.
@@ -24,13 +26,14 @@ The application allows controlling robotic mowers on a rectangular plateau, proc
 
 ## Assumptions & Decisions
 - Only one plateau is supported (as per MVP scope).
-- Input is received as a multiline string (via HTTP POST).
+- Input is received as a multiline string (via HTTP POST or CLI).
 - Directions are N, E, S, W. Actions are L (left), R (right), M (move forward).
 - If a mower tries to move to an invalid or occupied position, it stays in place.
 - The domain and service layers are pure Kotlin, with no Spring dependencies.
-- Parsers and mappers are placed in the web-layer for the MVP; for multi-interface support, move to a shared module.
+- Parsers and mappers are placed in the adapter-layer for reuse across interfaces.
 
 ## How to run
+### Web API
 1. **Build the project:**
    ```sh
    mvn clean install
@@ -43,26 +46,36 @@ The application allows controlling robotic mowers on a rectangular plateau, proc
 3. **Access Swagger UI:**
    - Open [http://localhost:8080/swagger-ui/index.html](http://localhost:8080/swagger-ui/index.html)
 
-4. **Example input:**
+### CLI
+NOTE: The CLI application needs JDK 17 or higher to run.
+1. **Build and run the CLI application:**
+   ```sh
+   sh ./cli-layer/run-cli.sh
+   # or, from inside cli-layer
+   sh ./run-cli.sh
    ```
-   5 5
-   1 2 N
-   LMLMLMLMM
-   3 3 E
-   MMRMMRMRRM
-   ```
-   **Expected output:**
-   ```
-   1 3 N
-   5 1 E
-   ```
+   - You will be prompted to enter the plateau size, mower positions, and actions interactively.
+
+## Example input (for both web and CLI)
+```
+5 5
+1 2 N
+LMLMLMLMM
+3 3 E
+MMRMMRMRRM
+```
+**Expected output:**
+```
+1 3 N
+5 1 E
+```
 
 ## Testing
 - Run all tests:
   ```sh
   mvn test
   ```
-- Includes unit tests (domain, service) and integration tests (controller, E2E)
+- Includes unit tests (domain, service, adapter) and integration tests (controller, E2E)
 
 ## Diagrams
 - See `docs/` for PlantUML diagrams of each layer.
