+++ adapter-layer/pom.xml
@@ -8,18 +8,21 @@
         <version>0.0.1-SNAPSHOT</version>
     </parent>
 
-    <artifactId>service-layer</artifactId>
+    <artifactId>adapter-layer</artifactId>
     <packaging>jar</packaging>
-    <name>service-layer</name>
-    <description>Business logic for MowerEngine</description>
+    <name>adapter-layer</name>
 
     <dependencies>
+        <dependency>
+            <groupId>com.seatcode.mowerengine</groupId>
+            <artifactId>service-layer</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <dependency>
             <groupId>com.seatcode.mowerengine</groupId>
             <artifactId>model-layer</artifactId>
             <version>${project.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
             <artifactId>kotlin-stdlib-jdk8</artifactId>
