+++ model-layer/pom.xml
@@ -2,20 +2,16 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
     <parent>
         <groupId>com.seatcode.mowerengine</groupId>
         <artifactId>mowerengine</artifactId>
         <version>0.0.1-SNAPSHOT</version>
     </parent>
+
     <artifactId>model-layer</artifactId>
     <packaging>jar</packaging>
     <name>model-layer</name>
     <description>Domain models for MowerEngine</description>
-
-    <properties>
-        <kotlin.version>1.9.24</kotlin.version>
-    </properties>
 
     <dependencies>
         <dependency>
@@ -44,21 +40,6 @@
             <plugin>
                 <groupId>org.jetbrains.kotlin</groupId>
                 <artifactId>kotlin-maven-plugin</artifactId>
-                <version>${kotlin.version}</version>
-                <executions>
-                    <execution>
-                        <id>compile</id>
-                        <goals>
-                            <goal>compile</goal>
-                        </goals>
-                    </execution>
-                    <execution>
-                        <id>test-compile</id>
-                        <goals>
-                            <goal>test-compile</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
         </plugins>
     </build>
