+++ pom.xml
@@ -4,15 +4,20 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.seatcode.mowerengine</groupId>
     <artifactId>mowerengine</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
     <packaging>pom</packaging>
+    <version>0.0.1-SNAPSHOT</version>
+
     <name>mowerengine</name>
     <description>MowerEngine Interview Project</description>
+
     <modules>
         <module>model-layer</module>
         <module>service-layer</module>
+        <module>adapter-layer</module>
         <module>web-layer</module>
+        <module>cli-layer</module>
     </modules>
+
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
@@ -26,6 +31,7 @@
         <junit.jupiter.version>5.10.2</junit.jupiter.version>
         <spring-boot.version>3.3.0</spring-boot.version>
     </properties>
+
     <dependencyManagement>
         <dependencies>
             <dependency>
@@ -53,6 +59,41 @@
                 <version>${junit.jupiter.version}</version>
                 <scope>test</scope>
             </dependency>
+            <dependency>
+                <groupId>org.jetbrains.kotlin</groupId>
+                <artifactId>kotlin-stdlib-jdk8</artifactId>
+                <version>${kotlin.version}</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
+
+    <build>
+        <sourceDirectory>src/main/kotlin</sourceDirectory>
+        <testSourceDirectory>src/test/kotlin</testSourceDirectory>
+        <pluginManagement>
+            <plugins>
+                <plugin>
+                    <groupId>org.jetbrains.kotlin</groupId>
+                    <artifactId>kotlin-maven-plugin</artifactId>
+                    <version>${kotlin.version}</version>
+                    <executions>
+                        <execution>
+                            <id>compile</id>
+                            <goals>
+                                <goal>compile</goal>
+                            </goals>
+                            <phase>process-sources</phase>
+                        </execution>
+                        <execution>
+                            <id>test-compile</id>
+                            <goals>
+                                <goal>test-compile</goal>
+                            </goals>
+                            <phase>process-test-sources</phase>
+                        </execution>
+                    </executions>
+                </plugin>
+            </plugins>
+        </pluginManagement>
+    </build>
 </project>
